[Refactor] dto 폴더 domain 폴더로 이동
